content-security-policy/script-src/script-src-strict_dynamic_parser_inserted.html

<!DOCTYPE HTML>
<html>

<head>
    <title>Parser-inserted scripts without a correct nonce are not allowed with `strict-dynamic` in the script-src directive.</title>
    <script src='/resources/testharness.js' nonce='dummy'></script>
    <script src='/resources/testharnessreport.js' nonce='dummy'></script>
    <script src='../support/checkSecurityPolicyViolations.sub.js' nonce='dummy'></script>
    <!-- CSP served: script-src 'strict-dynamic' 'nonce-dummy' -->
</head>

<body>
    <h1>Parser-inserted scripts without a correct nonce are not allowed with `strict-dynamic` in the script-src directive.</h1>
    <div id='log'></div>

    <script nonce='dummy'>
        var simpleSourcedScriptRan = false;
        async_test(function(t) {
            document.write('<scr' + 'ipt src="simpleSourcedScript.js"></scr' + 'ipt>');
            setTimeout(t.step_func_done(function() {
                assert_false(simpleSourcedScriptRan);
            }), 1000);
        }, 'Parser-inserted script via `document.write` without a correct nonce is not allowed with `strict-dynamic`.');
    </script>

    <script nonce='dummy'>
        var simpleSourcedScriptRan = false;
        async_test(function(t) {
            document.writeln('<scr' + 'ipt src="simpleSourcedScript.js"></scr' + 'ipt>');
            setTimeout(t.step_func_done(function() {
                assert_false(simpleSourcedScriptRan);
            }), 1000);
        }, 'Parser-inserted script via `document.writeln` without a correct nonce is not allowed with `strict-dynamic`.');
    </script>

    <script nonce='dummy'>
        var simpleSourcedScriptRan = false;
        async_test(function(t) {
            document.write('<scr' + 'ipt defer src="simpleSourcedScript.js"></scr' + 'ipt>');
            setTimeout(t.step_func_done(function() {
                assert_false(simpleSourcedScriptRan);
            }), 1000);
        }, 'Parser-inserted deferred script via `document.write` without a correct nonce is not allowed with `strict-dynamic`.');
    </script>

    <script nonce='dummy'>
        var simpleSourcedScriptRan = false;
        async_test(function(t) {
            document.writeln('<scr' + 'ipt defer src="simpleSourcedScript.js"></scr' + 'ipt>');
            setTimeout(t.step_func_done(function() {
                assert_false(simpleSourcedScriptRan);
            }), 1000);
        }, 'Parser-inserted deferred script via `document.writeln` without a correct nonce is not allowed with `strict-dynamic`.');
    </script>

    <script nonce='dummy'>
        var simpleSourcedScriptRan = false;
        async_test(function(t) {
            document.write('<scr' + 'ipt async src="simpleSourcedScript.js"></scr' + 'ipt>');
            setTimeout(t.step_func_done(function() {
                assert_false(simpleSourcedScriptRan);
            }), 1000);
        }, 'Parser-inserted async script via `document.write` without a correct nonce is not allowed with `strict-dynamic`.');
    </script>

    <script nonce='dummy'>
        var simpleSourcedScriptRan = false;
        async_test(function(t) {
            document.writeln('<scr' + 'ipt async src="simpleSourcedScript.js"></scr' + 'ipt>');
            setTimeout(t.step_func_done(function() {
                assert_false(simpleSourcedScriptRan);
            }), 1000);
        }, 'Parser-inserted async script via `document.writeln` without a correct nonce is not allowed with `strict-dynamic`.');
    </script>

    <script nonce='dummy'>
        var simpleSourcedScriptRan = false;
        async_test(function(t) {
            document.write('<scr' + 'ipt defer async src="simpleSourcedScript.js"></scr' + 'ipt>');
            setTimeout(t.step_func_done(function() {
                assert_false(simpleSourcedScriptRan);
            }), 1000);
        }, 'Parser-inserted deferred async script via `document.write` without a correct nonce is not allowed with `strict-dynamic`.');
    </script>

    <script nonce='dummy'>
        var simpleSourcedScriptRan = false;
        async_test(function(t) {
            document.writeln('<scr' + 'ipt defer async src="simpleSourcedScript.js"></scr' + 'ipt>');
            setTimeout(t.step_func_done(function() {
                assert_false(simpleSourcedScriptRan);
            }), 1000);
        }, 'Parser-inserted deferred async script via `document.writeln` without a correct nonce is not allowed with `strict-dynamic`.');
    </script>

    <script nonce='dummy'>
        var innerHTMLScriptRan = false;
        async_test(function(t) {
            var e = document.createElement('div');
            e.innerHTML = "<img src='/nonexisting.jpg' onerror='innerHTMLScriptRan = true'>";
            document.body.appendChild(e);
            setTimeout(t.step_func_done(function() {
                assert_false(innerHTMLScriptRan);
            }), 1000);
        }, 'Script injected via `innerHTML` is not allowed with `strict-dynamic`.');
    </script>

    <script nonce='dummy'>
        var insertAdjacentHTMLScriptRan = false;
        async_test(function(t) {
            var e = document.createElement('div');
            e.insertAdjacentHTML('afterbegin', "<img src='/nonexisting.jpg' onerror='insertAdjacentHTMLScriptRan = true'>");
            document.body.appendChild(e);
            setTimeout(t.step_func_done(function() {
                assert_false(insertAdjacentHTMLScriptRan);
            }), 1000);
        }, 'Script injected via `insertAdjacentHTML` is not allowed with `strict-dynamic`.');
    </script>

    <script nonce='dummy'>
        async_test(function(t) {
            setTimeout(t.step_func_done(function() {
                assert_violations({
                    'script-src': 10
                });
            }), 1000);
        }, 'All the CSP violation reports have been fired.');
    </script>

</body>

</html>
